cville-weather/requirements.txt
srequests==2.31.0
boto3==1.34.0
matplotlib==3.8.0
seaborn